requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
torch==2.2.1+cpu
torchvision==0.17.1+cpu
Pillow==10.2.0